Autogenerating page handlers, with reload buttons.
When you visit a location for the first time a page is auto-generated for
it including a "reload" button to trigger a POST request to the same
location/URL to regenerate the page.

There is also a facility to store per-page data for use in regenerating
the page.

The idea now is to allow for fields in the "reload" form/button so you
can edit the contents of a given page and POST them back to your server.