Factor out page import from main.
This means that all of the page rendering is in page.py while all of the
routing, request handling, and page caching is in main,py.  The last
little thing here that spells trouble is that the page rendering function
gets attached to the router as an attribute.  If I convert it into a
class and attach an instance instead I think the whole router object,
inluding the page handlers and the renderer instance, will be picklable.
;-)