Getting ready to integrate Joy into the system.
Also, stopped calling title() on the titles.  Edit them to be how you like.

Joy integration works like so: mention a link with the "command:" action
and it will be turned into a little form button with the title of the
link as it's text and a hidden field containing the link as its value.

Clicking on the link will cause the server to look up the linked page and
check its data for an 'expression' value, which will be a Joy expression.
Then this expression will be evaluated using the 'stack' value of the
page on which the command button is located (the same one that the button
posts to) and the result will become the new 'stack' value for the page.
(The 'stack' will default to an empty stack if it doesn't already exist.)