Ditch 'stack' in favor of 'joy'.
Now each page can have a 'joy' entry in the page_data that contains a
Joy expression.  If you use the page as the source of a command the
'joy' Joy expression is evaluated, and the 'joy' Joy expression of the
receiving page is used as the stack.

Basically the stack(s) and command expression(s) are now unified.

It is an experiment.